+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.5.1
+
+(#11)
+
+- Renable system font support.
+- Change font settings to be a text input for selective font loading.
+
+
 # 0.5.0
 - Update Typst version to 0.6.0
 - Attempt to fix bad canvas output